+++ src/main/java/blackjack/gambler/Players.java
@@ -1,12 +1,8 @@
 package blackjack.gambler;
 
 import blackjack.card.Cards;
-import blackjack.constant.MatchResult;
 import java.util.ArrayList;
-import java.util.LinkedHashMap;
 import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
 public class Players {
@@ -15,15 +11,6 @@ public class Players {
 
     public Players(List<Player> players) {
         this.players = new ArrayList<>(players);
-    }
-
-    public Map<Player, MatchResult> deriveResults(int dealerScore) {
-        return players.stream()
-                .collect(Collectors.toMap(
-                        player -> player,
-                        player -> player.compareTo(dealerScore),
-                        (x, y) -> y, LinkedHashMap::new
-                ));
     }
 
     public List<Player> getPlayers() {
